feat: #22 create searchbar
