+++ react-basic-practice/src/components/SideBar/index.css
@@ -1,0 +1,24 @@
+.sidebar.title {
+  font-size: var(--fs-large-plus);
+  font-weight: var(--fw-medium);
+}
+
+.search {
+  display: flex;
+  margin-top: 40px;
+  border-bottom: 1px solid var(--color-line);
+}
+
+.input-group {
+  padding: 0px;
+}
+
+.input-search {
+  border: none;
+  padding: 0px;
+}
+
+.btn-search {
+  padding: 0px;
+  background-color: var(--color-white);
+}
