+++ react-basic-practice/src/pages/HomePage/index.css
@@ -18,28 +18,3 @@
 .grid-container {
   justify-items: end;
 }
-
-.sidebar.title {
-  font-size: var(--fs-large-plus);
-  font-weight: var(--fw-medium);
-}
-
-.seach {
-  display: flex;
-  margin-top: 40px;
-  border-bottom: 1px solid var(--color-line);
-}
-
-.input-group {
-  padding: 0px;
-}
-
-.input-search {
-  border: none;
-  padding: 0px;
-}
-
-.btn-search {
-  padding: 0px;
-  background-color: var(--color-white);
-}
